requirements.txt
flask>=3.0.0
google-generativeai>=0.7.0
requests>=2.31.0
python-dotenv>=1.0.0
werkzeug>=3.0.0
Flask-SQLAlchemy>=3.1.1
